Add files via upload
If the length of the strings are different than each other its impossible to have valid anagram. If equals, creating two empty dictionaries to count the number of each occurences. 

I've tried to go with using set at the beginning but it doesn't work for such example:

s = "ccad" t = "caad"